requirements.txt
discord.py
Voice Support
gTTS
python-dotenv
FFmpeg	